подключает шаблоны в index.php
